mcp-autopilot/pyproject.toml
[project]
name = "mcp-autopilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp[cli]>=1.6.0",
    "PyYAML>=6.0.2"
]
